requirements.txt
streamlit==1.32.0
tensorflow==2.15.0
keras==2.15.0
numpy
pillow
opencv-python-headless
matplotlib
